Cichlid Mating Preference

Preference index in F1 and F2 crosses of two species of cichlids from Lake Victoria, Pundamilia pundamilia and P. nyererei.

Usage

Cichlids

Format

A data frame with 53 observations on the following 2 variables.

genotype

a factor with levels F1 and F2

preference

a numeric vector

Source

Haeslery, M.P. and O. Seehausen. 2005. Inheritance of female mating preference in a sympatric sibling species pair of Lake Victoria cichlids: implications for speciation. Proceedings of the Royal Society of London, Series B, Biological Sciences 272: 237-245.

References

http://rspb.royalsocietypublishing.org/content/272/1560/237.full.pdf

Examples

str(Cichlids)
#> 'data.frame':	53 obs. of  2 variables:
#>  $ genotype  : Factor w/ 2 levels "F1","F2": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ preference: num  0.108 0.098 0.078 0.072 0.062 0.047 0.007 0.007 0.007 0.007 ...

summary(preference ~ genotype, Cichlids, fun = favstats)
#>  Length   Class    Mode 
#>       3 formula    call 

if (require(plyr)) {
ddply(Cichlids, .(genotype),
  function(df)c(mean = mean(df$preference),
                standard.deviation = sd(df$preference),
                n = length(df$preference)))
}
#> Loading required package: plyr
#> ------------------------------------------------------------------------------
#> You have loaded plyr after dplyr - this is likely to cause problems.
#> If you need functions from both plyr and dplyr, please load plyr first, then dplyr:
#> library(plyr); library(dplyr)
#> ------------------------------------------------------------------------------
#> 
#> Attaching package: ‘plyr’
#> The following object is masked from ‘package:mosaic’:
#> 
#>     count
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     arrange, count, desc, failwith, id, mutate, rename, summarise,
#>     summarize
#>   genotype        mean standard.deviation  n
#> 1       F1 0.004900000         0.06424288 20
#> 2       F2 0.008151515         0.15820460 33