Convictions and Income Level in a Cohort of English Boys

Data reporting the number of individual with and without convictions per income level.

Usage

ConvictionsAndIncome

Format

A data frame with 395 observations on the following 2 variables.

convicted

a factor with levels no and yes

income

a factor with levels adequate, comfortable, and inadequate

Source

Farrington, D.P. 1994. Cambridge Study in Delinquent Development [Great Britain], 1961-1981. 2nd ICPSR ed. Inter-university Consortium for Political and Social Research, Ann Arbor, MI.

References

http://www.icpsr.umich.edu/icpsrweb/NACJD/archive.jsp

Examples

str(ConvictionsAndIncome)
#> 'data.frame':	395 obs. of  2 variables:
#>  $ convicted: Factor w/ 2 levels "no","yes": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ income   : Factor w/ 3 levels "adequate","comfortable",..: 2 2 2 2 2 2 2 2 2 2 ...
ConvictionsAndIncome
#>     convicted      income
#> 1         yes comfortable
#> 2         yes comfortable
#> 3         yes comfortable
#> 4         yes comfortable
#> 5         yes comfortable
#> 6         yes comfortable
#> 7         yes comfortable
#> 8         yes comfortable
#> 9         yes comfortable
#> 10        yes comfortable
#> 11        yes comfortable
#> 12        yes comfortable
#> 13        yes comfortable
#> 14        yes comfortable
#> 15        yes comfortable
#> 16        yes comfortable
#> 17        yes comfortable
#> 18        yes comfortable
#> 19        yes comfortable
#> 20        yes comfortable
#> 21        yes comfortable
#> 22        yes comfortable
#> 23        yes comfortable
#> 24        yes comfortable
#> 25        yes comfortable
#> 26        yes comfortable
#> 27        yes comfortable
#> 28        yes comfortable
#> 29        yes comfortable
#> 30        yes comfortable
#> 31        yes    adequate
#> 32        yes    adequate
#> 33        yes    adequate
#> 34        yes    adequate
#> 35        yes    adequate
#> 36        yes    adequate
#> 37        yes    adequate
#> 38        yes    adequate
#> 39        yes    adequate
#> 40        yes    adequate
#> 41        yes    adequate
#> 42        yes    adequate
#> 43        yes    adequate
#> 44        yes    adequate
#> 45        yes    adequate
#> 46        yes    adequate
#> 47        yes    adequate
#> 48        yes    adequate
#> 49        yes    adequate
#> 50        yes    adequate
#> 51        yes    adequate
#> 52        yes    adequate
#> 53        yes    adequate
#> 54        yes    adequate
#> 55        yes    adequate
#> 56        yes    adequate
#> 57        yes    adequate
#> 58        yes    adequate
#> 59        yes    adequate
#> 60        yes    adequate
#> 61        yes    adequate
#> 62        yes    adequate
#> 63        yes    adequate
#> 64        yes    adequate
#> 65        yes    adequate
#> 66        yes    adequate
#> 67        yes    adequate
#> 68        yes    adequate
#> 69        yes    adequate
#> 70        yes    adequate
#> 71        yes    adequate
#> 72        yes    adequate
#> 73        yes    adequate
#> 74        yes    adequate
#> 75        yes    adequate
#> 76        yes    adequate
#> 77        yes    adequate
#> 78        yes    adequate
#> 79        yes    adequate
#> 80        yes    adequate
#> 81        yes    adequate
#> 82        yes    adequate
#> 83        yes    adequate
#> 84        yes    adequate
#> 85        yes    adequate
#> 86        yes    adequate
#> 87        yes    adequate
#> 88        yes  inadequate
#> 89        yes  inadequate
#> 90        yes  inadequate
#> 91        yes  inadequate
#> 92        yes  inadequate
#> 93        yes  inadequate
#> 94        yes  inadequate
#> 95        yes  inadequate
#> 96        yes  inadequate
#> 97        yes  inadequate
#> 98        yes  inadequate
#> 99        yes  inadequate
#> 100       yes  inadequate
#> 101       yes  inadequate
#> 102       yes  inadequate
#> 103       yes  inadequate
#> 104       yes  inadequate
#> 105       yes  inadequate
#> 106       yes  inadequate
#> 107       yes  inadequate
#> 108       yes  inadequate
#> 109       yes  inadequate
#> 110       yes  inadequate
#> 111       yes  inadequate
#> 112       yes  inadequate
#> 113       yes  inadequate
#> 114       yes  inadequate
#> 115       yes  inadequate
#> 116       yes  inadequate
#> 117       yes  inadequate
#> 118       yes  inadequate
#> 119       yes  inadequate
#> 120       yes  inadequate
#> 121       yes  inadequate
#> 122       yes  inadequate
#> 123       yes  inadequate
#> 124       yes  inadequate
#> 125       yes  inadequate
#> 126       yes  inadequate
#> 127       yes  inadequate
#> 128       yes  inadequate
#> 129       yes  inadequate
#> 130       yes  inadequate
#> 131        no comfortable
#> 132        no comfortable
#> 133        no comfortable
#> 134        no comfortable
#> 135        no comfortable
#> 136        no comfortable
#> 137        no comfortable
#> 138        no comfortable
#> 139        no comfortable
#> 140        no comfortable
#> 141        no comfortable
#> 142        no comfortable
#> 143        no comfortable
#> 144        no comfortable
#> 145        no comfortable
#> 146        no comfortable
#> 147        no comfortable
#> 148        no comfortable
#> 149        no comfortable
#> 150        no comfortable
#> 151        no comfortable
#> 152        no comfortable
#> 153        no comfortable
#> 154        no comfortable
#> 155        no comfortable
#> 156        no comfortable
#> 157        no comfortable
#> 158        no comfortable
#> 159        no comfortable
#> 160        no comfortable
#> 161        no comfortable
#> 162        no comfortable
#> 163        no comfortable
#> 164        no comfortable
#> 165        no comfortable
#> 166        no comfortable
#> 167        no comfortable
#> 168        no comfortable
#> 169        no comfortable
#> 170        no comfortable
#> 171        no comfortable
#> 172        no comfortable
#> 173        no comfortable
#> 174        no comfortable
#> 175        no comfortable
#> 176        no comfortable
#> 177        no comfortable
#> 178        no comfortable
#> 179        no comfortable
#> 180        no comfortable
#> 181        no comfortable
#> 182        no comfortable
#> 183        no comfortable
#> 184        no comfortable
#> 185        no comfortable
#> 186        no comfortable
#> 187        no comfortable
#> 188        no comfortable
#> 189        no comfortable
#> 190        no comfortable
#> 191        no comfortable
#> 192        no comfortable
#> 193        no comfortable
#> 194        no comfortable
#> 195        no comfortable
#> 196        no comfortable
#> 197        no comfortable
#> 198        no comfortable
#> 199        no comfortable
#> 200        no comfortable
#> 201        no comfortable
#> 202        no comfortable
#> 203        no comfortable
#> 204        no comfortable
#> 205        no comfortable
#> 206        no comfortable
#> 207        no comfortable
#> 208        no comfortable
#> 209        no comfortable
#> 210        no comfortable
#> 211        no comfortable
#> 212        no comfortable
#> 213        no comfortable
#> 214        no comfortable
#> 215        no comfortable
#> 216        no comfortable
#> 217        no comfortable
#> 218        no comfortable
#> 219        no comfortable
#> 220        no comfortable
#> 221        no    adequate
#> 222        no    adequate
#> 223        no    adequate
#> 224        no    adequate
#> 225        no    adequate
#> 226        no    adequate
#> 227        no    adequate
#> 228        no    adequate
#> 229        no    adequate
#> 230        no    adequate
#> 231        no    adequate
#> 232        no    adequate
#> 233        no    adequate
#> 234        no    adequate
#> 235        no    adequate
#> 236        no    adequate
#> 237        no    adequate
#> 238        no    adequate
#> 239        no    adequate
#> 240        no    adequate
#> 241        no    adequate
#> 242        no    adequate
#> 243        no    adequate
#> 244        no    adequate
#> 245        no    adequate
#> 246        no    adequate
#> 247        no    adequate
#> 248        no    adequate
#> 249        no    adequate
#> 250        no    adequate
#> 251        no    adequate
#> 252        no    adequate
#> 253        no    adequate
#> 254        no    adequate
#> 255        no    adequate
#> 256        no    adequate
#> 257        no    adequate
#> 258        no    adequate
#> 259        no    adequate
#> 260        no    adequate
#> 261        no    adequate
#> 262        no    adequate
#> 263        no    adequate
#> 264        no    adequate
#> 265        no    adequate
#> 266        no    adequate
#> 267        no    adequate
#> 268        no    adequate
#> 269        no    adequate
#> 270        no    adequate
#> 271        no    adequate
#> 272        no    adequate
#> 273        no    adequate
#> 274        no    adequate
#> 275        no    adequate
#> 276        no    adequate
#> 277        no    adequate
#> 278        no    adequate
#> 279        no    adequate
#> 280        no    adequate
#> 281        no    adequate
#> 282        no    adequate
#> 283        no    adequate
#> 284        no    adequate
#> 285        no    adequate
#> 286        no    adequate
#> 287        no    adequate
#> 288        no    adequate
#> 289        no    adequate
#> 290        no    adequate
#> 291        no    adequate
#> 292        no    adequate
#> 293        no    adequate
#> 294        no    adequate
#> 295        no    adequate
#> 296        no    adequate
#> 297        no    adequate
#> 298        no    adequate
#> 299        no    adequate
#> 300        no    adequate
#> 301        no    adequate
#> 302        no    adequate
#> 303        no    adequate
#> 304        no    adequate
#> 305        no    adequate
#> 306        no    adequate
#> 307        no    adequate
#> 308        no    adequate
#> 309        no    adequate
#> 310        no    adequate
#> 311        no    adequate
#> 312        no    adequate
#> 313        no    adequate
#> 314        no    adequate
#> 315        no    adequate
#> 316        no    adequate
#> 317        no    adequate
#> 318        no    adequate
#> 319        no    adequate
#> 320        no    adequate
#> 321        no    adequate
#> 322        no    adequate
#> 323        no    adequate
#> 324        no    adequate
#> 325        no    adequate
#> 326        no    adequate
#> 327        no    adequate
#> 328        no    adequate
#> 329        no    adequate
#> 330        no    adequate
#> 331        no    adequate
#> 332        no    adequate
#> 333        no    adequate
#> 334        no    adequate
#> 335        no    adequate
#> 336        no    adequate
#> 337        no    adequate
#> 338        no    adequate
#> 339        no    adequate
#> 340        no    adequate
#> 341        no    adequate
#> 342        no    adequate
#> 343        no    adequate
#> 344        no    adequate
#> 345        no    adequate
#> 346        no    adequate
#> 347        no    adequate
#> 348        no    adequate
#> 349        no  inadequate
#> 350        no  inadequate
#> 351        no  inadequate
#> 352        no  inadequate
#> 353        no  inadequate
#> 354        no  inadequate
#> 355        no  inadequate
#> 356        no  inadequate
#> 357        no  inadequate
#> 358        no  inadequate
#> 359        no  inadequate
#> 360        no  inadequate
#> 361        no  inadequate
#> 362        no  inadequate
#> 363        no  inadequate
#> 364        no  inadequate
#> 365        no  inadequate
#> 366        no  inadequate
#> 367        no  inadequate
#> 368        no  inadequate
#> 369        no  inadequate
#> 370        no  inadequate
#> 371        no  inadequate
#> 372        no  inadequate
#> 373        no  inadequate
#> 374        no  inadequate
#> 375        no  inadequate
#> 376        no  inadequate
#> 377        no  inadequate
#> 378        no  inadequate
#> 379        no  inadequate
#> 380        no  inadequate
#> 381        no  inadequate
#> 382        no  inadequate
#> 383        no  inadequate
#> 384        no  inadequate
#> 385        no  inadequate
#> 386        no  inadequate
#> 387        no  inadequate
#> 388        no  inadequate
#> 389        no  inadequate
#> 390        no  inadequate
#> 391        no  inadequate
#> 392        no  inadequate
#> 393        no  inadequate
#> 394        no  inadequate
#> 395        no  inadequate

xtabs(~ convicted + income, data = ConvictionsAndIncome)
#>          income
#> convicted adequate comfortable inadequate
#>       no       128          90         47
#>       yes       57          30         43