Effects of Trimethadione on Lifespan in Nematodes

lifespan of the nematode Caenorhabditis elegans in control and three experimental treatments of the anticonvulsant drug trimethadione.

Format

A data frame with 200 observations on the following 2 variables.

treatment

a factor with levels: adult, larva, larva+adult, and water

lifespan

a numeric vector of lifespan

Source

inferred from Evason, K., C. Huang, I. Yamben, D.F. Covey, and K. Kornfeld. 2005. Anticonvulsant medications extend worm life-span. Science 307: 258-262.

References

http://www.sciencemag.org/cgi/content/abstract/307/5707/258

Examples

str(NematodeLifespan)
#> 'data.frame':	200 obs. of  2 variables:
#>  $ treatment: Factor w/ 4 levels "adult","larva",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ lifespan : int  6 10 12 12 13 13 14 14 14 14 ...