Frequencies of Fish Eaten by Trematode Infection Level

Frequencies of killifish (Fundulus parvipinnis) eaten by birds depending on level of infection by the trematode Euhaplorchis californiensis.

Usage

Trematodes

Format

A data frame with 141 observations on the following 2 variables.

infection.status

a factor with levels: high, light, and uninfected

eaten

a factor with levels: no and yes

Source

Lafferty, K.D. and A.K. Morris. 1996. Altered behavior of parasitized killifish increases susceptibility to predation by bird final hosts. Ecology 77: 1390-1397.

Examples

demo(sec9.3)
#> 
#> 
#> 	demo(sec9.3)
#> 	---- ~~~~~~
#> 
#> > ## Table 9.3-1
#> > xtabs(~ infection.status + eaten, Trematodes)
#>                 eaten
#> infection.status no yes
#>       high        9  37
#>       light      35  10
#>       uninfected 49   1
#> 
#> > ## Chi-squared Contingency Test
#> > chisq.test( xtabs(~ infection.status + eaten, Trematodes) )
#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  xtabs(~infection.status + eaten, Trematodes)
#> X-squared = 69.756, df = 2, p-value = 7.124e-16
#> 
#> 
#> > summary(chisq.test( xtabs(~ infection.status + eaten, Trematodes) ) )
#>           Length Class  Mode     
#> statistic 1      -none- numeric  
#> parameter 1      -none- numeric  
#> p.value   1      -none- numeric  
#> method    1      -none- character
#> data.name 1      -none- character
#> observed  6      xtabs  numeric  
#> expected  6      -none- numeric  
#> residuals 6      xtabs  numeric  
#> stdres    6      xtabs  numeric